rename header chat button to chat header button
